db_api/requirements.txt
psycopg2
Flask
Pillow